Update `UserId` type to `long`, add presigned URL retrieval in DigitalOcean storage service, allow anonymous access to image profile update, and adjust `GetUserAsyncById` to support `long`.

+++ CryptoJackpotService.Data/Repositories/IRepositories/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace CryptoJackpotService.Data.Repositories.IRepositories;
 public interface IUserRepository
 { 
     Task<User> CreateUserAsync(User user);
-    Task<User?> GetUserAsyncById(int id);
+    Task<User?> GetUserAsyncById(long id);
     Task<User?> GetUserAsyncByEmail(string email);
     Task<User> UpdateUserAsync(User user);
 }
